Complete refactor and rewrite
Major update to refactor the class so it only contains customizations of the standard book class. Standard book class options are now passed to the book class and and the book class is loaded so that anything not handled by our class falls back to the book class.

We also added options for minimal styling (uses a minimum set of custom styling), no styling (no custom styling everything is styled as in the book class), and single spacing.